CSS Got messed up
I figured out what was going on. I duplicated the files and put them into a new folder to organize for homework 18 and didn't update the path name of the css sheet in the html. So now CSS is working, still need to tweak now that I can see live CSS